Finalized minor changes to the CRUD operations

+++ app/Http/Controllers/InfluencerCategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\InfluencerCategory;
 use Illuminate\Http\Request;
 
+
 class InfluencerCategoryController extends Controller
 {
     public function store(Request $request)
@@ -34,9 +35,18 @@ class InfluencerCategoryController extends Controller
 
     public function destroy(InfluencerCategory $influencerCategory)
     {
-        $influencerCategory->delete();
+        // Check if there are any InfluencerCards associated with the InfluencerCategory
+        if ($influencerCategory->influencerCards()->exists()) {
+            // If there are, return an error message and stop the deletion process
+            return redirect()->route('category.index')
+                ->with('error', 'Cannot delete this Influencer Category as it is being used by some Influencer Cards');
+        } else {
+            // If there are no InfluencerCards associated with the InfluencerCategory, delete it
+            $influencerCategory->delete();
 
-        return redirect()->route('category.index')
-            ->with('success', 'Influencer Category deleted successfully');
+            return redirect()->route('category.index')
+                ->with('success', 'Influencer Category deleted successfully');
+        }
     }
+
 }
